Break out JPanels into seperate classes and add callbacks
