+++ CS245Project1.iml
@@ -9,4 +9,3 @@
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
-
