+++ src/callbacks/FrameCallbacks.java
@@ -1,0 +1,13 @@
+package callbacks;
+
+/**
+ * Created by cary on 1/13/16.
+ */
+public interface FrameCallbacks {
+
+    void startMainMenu();
+
+    void startSplashScreen();
+
+    void startNewGame();
+}
