+++ src/game/HangmanGame.java
@@ -1,0 +1,7 @@
+package game;
+
+/**
+ * Created by cary on 1/13/16.
+ */
+public class HangmanGame {
+}
